Nuevo método CreateMarbles y cambio de tipo de variable

+++ DeathTokenBase/DeathToken/Button.cpp
@@ -42,10 +42,10 @@ void Button::connect(Callback callback) {
 
 ButtonUI::ButtonUI(GameState* g, int x, int y, int w, int h, Texture* t, Texture* tC)
 	: Button(g, x, y, w, h, t), textC(tC), clicked(false) {
-	boxB.x = x - (w * 0.05f);
-	boxB.y = y - (h * 0.05f);
-	boxB.w = w * 1.1f;
-	boxB.h = h * 1.1f;
+	boxB.x =(int)( x - (w * 0.05f));
+	boxB.y = (int)(y - (h * 0.05f));
+	boxB.w =(int)( w * 1.1f);
+	boxB.h = (int)(h * 1.1f);
 }
 void
 ButtonUI::update()
@@ -164,10 +164,10 @@ ButtonMarbles::render() const
 	case 1: {
 		for (int i = 0; i < NCMarbles.size(); i++) {
 			if (NCMarbles[i] == 1) {
-				auxBox.x = box.x + box.w / 2 - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-				auxBox.y = box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-				auxBox.w = (74.0 / 1920.0 * Game::WIN_WIDTH);
-				auxBox.h = (74.0 / 1080.0 * Game::WIN_HEIGHT);
+				auxBox.x = (int)(box.x + box.w / 2.0 - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2.0);
+				auxBox.y = (int)(box.y + box.h / 2.0 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2.0);
+				auxBox.w = (int)((74.0 / 1920.0 * Game::WIN_WIDTH));
+				auxBox.h = (int)((74.0 / 1080.0 * Game::WIN_HEIGHT));
 				CMarbles[i]->render(auxBox);
 				break;
 
@@ -179,36 +179,36 @@ ButtonMarbles::render() const
 		for (int i = 0; i < NCMarbles.size(); i++) {
 			if (NCMarbles[i] == 2) {
 
-				auxBox.x = box.x + box.w / 4 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-				auxBox.y = box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-				auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-				auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+				auxBox.x = (int)(box.x + box.w / 4.0 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+				auxBox.y = (int)(box.y + box.h / 2.0 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+				auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+				auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 				CMarbles[i]->render(auxBox);
 
 
 
-				auxBox.x = box.x + (box.w - box.w / 4) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-				auxBox.y = box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-				auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-				auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+				auxBox.x = (int)(box.x + (box.w - box.w / 4) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+				auxBox.y = (int)(box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+				auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+				auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 				CMarbles[i]->render(auxBox);
 
 
 			}
 			else if (NCMarbles[i] == 1) {
 				if (pos == 0) {
-					auxBox.x = box.x + box.w / 4 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + box.w / 4 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					pos++;
 				}
 				else {
-					auxBox.x = box.x + (box.w - box.w / 4) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + (box.w - box.w / 4) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + box.h / 2 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					break;
 				}
@@ -218,20 +218,20 @@ ButtonMarbles::render() const
 	}case 3: {
 		for (int i = 0; i < NCMarbles.size(); i++) {
 			if (NCMarbles[i] == 3) {
-				auxBox.x = box.x + box.w / 4 - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-				auxBox.y = box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-				auxBox.w = (74.0 / 1920.0 * Game::WIN_WIDTH);
-				auxBox.h = (74.0 / 1080.0 * Game::WIN_HEIGHT);
+				auxBox.x = (int)(box.x + box.w / 4 - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+				auxBox.y = (int)(box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+				auxBox.w = (int)(74.0 / 1920.0 * Game::WIN_WIDTH);
+				auxBox.h = (int)(74.0 / 1080.0 * Game::WIN_HEIGHT);
 				CMarbles[i]->render(auxBox);
-				auxBox.x = box.x + (box.w - box.w / 4) - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-				auxBox.y = box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-				auxBox.w = (74.0 / 1920.0 * Game::WIN_WIDTH);
-				auxBox.h = (74.0 / 1080.0 * Game::WIN_HEIGHT);
+				auxBox.x = (int)(box.x + (box.w - box.w / 4) - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+				auxBox.y = (int)(box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+				auxBox.w = (int)(74.0 / 1920.0 * Game::WIN_WIDTH);
+				auxBox.h = (int)(74.0 / 1080.0 * Game::WIN_HEIGHT);
 				CMarbles[i]->render(auxBox);
-				auxBox.x = box.x + 2 * box.w / 4 - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-				auxBox.y = box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-				auxBox.w = (74.0 / 1920.0 * Game::WIN_WIDTH);
-				auxBox.h = (74.0 / 1080.0 * Game::WIN_HEIGHT);
+				auxBox.x = (int)(box.x + 2 * box.w / 4 - (74.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+				auxBox.y = (int)(box.y + box.h / 2 - (74.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+				auxBox.w = (int)(74.0 / 1920.0 * Game::WIN_WIDTH);
+				auxBox.h = (int)(74.0 / 1080.0 * Game::WIN_HEIGHT);
 				CMarbles[i]->render(auxBox);
 				break;
 			}
@@ -243,88 +243,88 @@ ButtonMarbles::render() const
 			if (NCMarbles[i] == 3) {
 				if (i == 0) {
 					//1/1
-					auxBox.x = box.x + box.w / 5 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + box.w / 5 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//2/1
-					auxBox.x = box.x + box.w / 5 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + box.w / 5 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//3/1
-					auxBox.x = box.x + box.w / 5 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + box.w / 5 - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 
 
 				}
 				else if (i == 1) {
 					//1/2
-					auxBox.x = box.x + 2 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 2 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//2/2
-					auxBox.x = box.x + 2 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 2 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//3/2
-					auxBox.x = box.x + 2 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 2 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 
 
 				}
 				else if (i == 2) {
 					//1/3
-					auxBox.x = box.x + 3 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 3 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//2/3
-					auxBox.x = box.x + 3 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 3 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//3/3
-					auxBox.x = box.x + 3 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 3 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 
 				}
 				else if (i == 3) {
 
 					//1/4
-					auxBox.x = box.x + 4 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 4 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + box.h / 4 - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//2/4
-					auxBox.x = box.x + 4 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 4 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 2 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 					//3/4
-					auxBox.x = box.x + 4 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2;
-					auxBox.y = box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2;
-					auxBox.w = (52.0 / 1920.0 * Game::WIN_WIDTH);
-					auxBox.h = (52.0 / 1080.0 * Game::WIN_HEIGHT);
+					auxBox.x = (int)(box.x + 4 * (box.w / 5) - (52.0 / 1080.0 * Game::WIN_HEIGHT) / 2);
+					auxBox.y = (int)(box.y + 3 * (box.h / 4) - (52.0 / 1920.0 * Game::WIN_WIDTH) / 2);
+					auxBox.w = (int)(52.0 / 1920.0 * Game::WIN_WIDTH);
+					auxBox.h = (int)(52.0 / 1080.0 * Game::WIN_HEIGHT);
 					CMarbles[i]->render(auxBox);
 				}
 			}
